// src/app/models/baseball-card.model.ts
export interface BaseballCard {
    id?: any,
    inventoryId?: string;
    year: string;
    brand: string;
    cardNumber: string;
    grade: string;
    firstName: string;
    lastName: string;
    buyPrice?: string;
    buyDate?: string;
    sellingPrice?: string;
    soldPrice?: string;
    soldDate?: string;
}